[buildrng] Implemented a basic page linearizer for better RNG
The linearization code in linear.xsl should move to yelp-xsl some time,
because there are a bunch of uses for it.  But it should also be cleaned
up.  For now, this linearization allows us to produce RNC and RNG with a
better order without having to hard-code a page ordering.