[1.0/mal_page.page] Changing page/@id to xsd:NMTOKEN instead of xsd:ID
Sure, it's an ID.  But it only needs to be unique to other page IDs, which
are in other XML documents, so outside the scope of what RNG validation can
do.  Using xsd:ID only prevents page IDs from being the same as section IDs,
and that's not really a constraint we need to enforce in Mallard.