ducktype/1.0: Comments don't take priority over fences
I've gone back and forth on this. Advantages of comment priority is that you
can easily grep (or grep out) comments, and if you need to comment out stuff
in a fence, you can do that without extra fence markers. The advantage fence
priority is that you can reliably stuff anything (including Ducktype) in
fences and not worry, as long as there are no standalone ]]] lines.

In the end, I went with fence priority because it matches XML's priority of
CDATA over comments.